refactor due to review comments, extract unlock_conditions into separate file
