+++ crates/iota-types/src/stardust/output/mod.rs
@@ -2,6 +2,7 @@ pub mod alias;
 pub mod basic;
 pub mod foundry;
 pub mod nft;
+pub mod unlock_conditions;
 
 pub use alias::*;
 pub use basic::*;
